test(authz): drive invitation leak test through the real invite endpoint
Replace the direct ensureMemberEntity unit test with an integration test that
POSTs /api/auth/organization/invite-member and exercises the actual
afterCreateInvitation hook — so it regresses if the hook is re-broken even while
ensureMemberEntity stays correct. Verified red (inviter claim length 1) with the
buggy userId:inviter.id call, green with createdByUserId. Also adopt EntityData
typing in ensureMemberEntity, dropping the as-any casts.
